+++ CHANGELOG.md
@@ -6,7 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-Nothing here.
+Nothing yet.
+
+## [2.0.0] 2023-02-22
+
+### Changed
+
+- Added length limit on AggregateId values.
+- Refactored event store methods.
 
 ## [1.2.2] - 2023-02-22
 
@@ -47,6 +54,7 @@ Nothing here.
 - Implemented the EntityFrameworkCore.PostgreSQL event store.
 - Configured NuGet packages.
 
-[unreleased]: https://github.com/Logitar/EventSourcing/compare/v1.2.2...HEAD
+[unreleased]: https://github.com/Logitar/EventSourcing/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/Logitar/EventSourcing/compare/v1.2.2...v2.0.0
 [1.2.2]: https://github.com/Logitar/EventSourcing/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/Logitar/EventSourcing/releases/tag/v1.2.1